+++ regression/reference/templates/wrapstd_vector_int.h
@@ -33,7 +33,9 @@ extern "C" {
 #endif
 
 // typedef std::vector<int>::size_type
+// splicer begin namespace.std.class.vector.typedef.size_type
 typedef size_t TEM_vector_int_size_type;
+// splicer end namespace.std.class.vector.typedef.size_type
 
 // splicer begin namespace.std.class.vector.C_declarations
 // splicer end namespace.std.class.vector.C_declarations